Hook up the handler of the button and make available the url in it
The url is public on purpose in the repo because I wanted to share the
google sheet anyway.

Also removed the form tag since we are not using it for anything.

Also to prevent the button to defaulting to a submit button, because we
don't want the default submit to trigger but to handle it ourselves.

If we wanted the form tag anyways (for css styling for example) we could
always use `event.preventDefault()`.